Add Host Reservation Admin API
Initial release of Host Reservation Admin API by adding new classes
to the current Lease API mapped to /os-hosts/ context.

The goal of this API is to be admin-only, defining CRUD methods for
adding a host to Compute, aiming to dedicate it for lease management.

As a result, the added host won't be available for booting VMs on it
unless Climate specifically picks it from a specific freepool to the
user' pcloud associated with the host reservation.

For testing, please patch RPC methods like this :
    call = lambda *args, **kwargs: True
    cast = lambda *args, **kwargs: True

TODO :
 - Add @policy.authorize() to climate.api.oshosts.service controllers
   once review 57200 (Policy mgmt) is merged

Implements bp:host-provisioning-api

Change-Id: I58d986fe8344b9578b2f15399ec19f7649cc3035