Support for DB migrations using Alembic
Added support for DB migration using Alembic.
Modified devstack/lib/climate to sync the database during the
configuration of Climate.

Added the actual state of the sqlalchemy models in a migration script.
This will be the first version, and will be used by devstack and the
migration tests.

Added a new console_script to the setup.cfg: climate-db-manage. This is
the CLI is used as a wrapper of the alembic functionality.

Added alembic>=0.4.1 as dependecy.

Added some README with documentation about the DB migrations.

Change-Id: I390ccfac1e436db0b04339e60f9f6795b22b8f7e
Implements: blueprint schema-data-migration-with-alembic