Fix work with new keystoneclient master
When creating a keystoneclient connection with a trust, we should
not pass the project_name, as the token can only be scoped to either
a trust or project, not both.  Previous versions of keystoneclient
worked with both, but recent changes mean this no longer works.

Here is the keystoneclient commit that changed it:
https://github.com/openstack/python-keystoneclient/commit/7f1881211da4c22a718223af80645d5bfb609f3d

Change-Id: I0f6f68405bda69bc5d5194c6bae80d392a0714fb
Closes-Bug: #1289295