Climate Manager now starts even if a plugin fails
The initialization of plugins loaded by Climate Manager was deferred in
order to catch any exception that the plugins could raise during it
initialization. Doing this, Climate Manager starts correctly loading
only the plugins that could be correctly initialized. A warning is
logged when this situation happens.
When a lease is about to be created, the resource_type of any
reservation that could contain, is checked against all the registered
plugins. If none of them handles the resource_type, an
UnsupportedResourceType exception is raised.

Change-Id: I356a39d432dc2cdaf349d817d8b341c531ece5b0
Closes-Bug: #1294175