Add scenario testing support
Adding infrastructure to support standalone scenario testing
for Climate.

To make it happen, we had to mimic the way that tempest was
handling the scenario tests and since we weren't allow to modify
tempest code, we built a similar hierachy (check manager and scenario
classes and their parents).
As long for the config_resource_reservation file, we added some
params to handle the timeout checks

The test added follow these steps:
    1) Create an instance with the hint parameters
      * start_date: now + 1 minute
      * end_date: start_date + 3 minutes
    2) check vm was shelved
    3) check vm became active
    4) check that a new lease is created on climate
    5) check its param
    6) wait lease end
    7) make sure VM was snapshoted and removed
Test duration: between 3 and 5 minutes

Partially implements: blueprint climate-tempest-testing
Change-Id: I023866ec8f19cfa8fec16c8fb9af4ed014160fe0