Set osapi_compute_extensions default value
The Devstack commit 3317437caca5eb769ab4c6d3ea7c5d3e062517e2 removed
the osapi_compute_extensions configuration option from lib/nova.
This caused that when our Devstack implementation sets this
configuration option, the default value were missed. This default value
is important because is the one that loads all the default Nova
extensions, which causes a lot of HTTP 404 at Nova API.
For this reason, our Devstack implementation was modified in order to
set this default value when this option is not set by any other project.

Change-Id: I04698643ac00b4b52b05f28a6c46522991734fc4
Closes-Bug: #1312368