Passing auth_url to keystoneclient.authenticate
During the creation of a trust, our Keystone client implementation calls
the authenticate method which fails due to the auth_url parameter is
missing. Seems that the current V3 Kestone client isn't honoring the auth_url
passed during initialization.
If you call the authenticate method without passing the auth_url
parameter, the authentication fails because that parameter was not
properly set, but should work because that parameter is passed during
initialization.
This commit tries to fix the issue until a patch from Keystone lands to
master.

Change-Id: Id242dcac31fe535836033dff951844a66bcc2100
Closes-Bug: #1326066