Updated scenario tests to latest Tempest master
The __init__ method of the OfficialClientManager Tempest class changed
since we implemented our Tempest test. Now this method receives only one
parameter with all the credentials instead of three (user, password,
project). To fix this the ResourceReservationScenarioTest was modified
to use the new API. Also a not valid import was fixed and some minor
things of the test were fixed.

Change-Id: Ic79bdfb0e85d7a47310978a5c3ebd6a37195795e
Closes-Bug: #1326532