Update method in host_plugin to find hypervisor by name or id
The on_end event for physical reservations was throwing an error as the
host_plugin was trying to find hypervisor/host by name using an ID.

Updated the method to lookup hypervisor from the hypervisor list using
either the name or id. Also updated other references in the code and the
tests.

Change-Id: I6d5a0fea5beeb2e62aad1954591f6e1ced1aceaf
Closes-Bug: 1333342