Use oslo.serialization
This allows to remove the wsgi and sslutils code originating from
oslo-incubator. sslutils is causing ubuntu-xenial gate jobs to fail
because it refers to ssl.PROTOCOL_SSLv3, which was disabled in OpenSSL
on Debian and Ubuntu due to the POODLE attack.

Also update climate.conf sample file to pick up library changes in this
commit and previous ones. This required patching the generator code to
support cfg.PortOpt and cfg.Opt, but ultimately will be replaced to use
the new generator in oslo.config.

Change-Id: I54fcc1cf6c86bbadf6e922f7339d2bde18182e3a