Enable DevStack to set Keystone service name in config
With default DevStack settings, the Keystone service name is defined as
'identity'. However, Blazar expects it to be 'identityv3' by default. It
is used to get the corresponding Keystone endpoint from the service
catalog.

This patch makes the value configurable, which Jenkins jobs can use to
set the Keystone service name.

Change-Id: I9fdd4207d31364697e9721db5ea96ccc34efbbe7