Use oslo-config-generator
This commit replaces the old configuration file generator from the
openstack.common.config module by oslo-config-generator [1]. It also
deletes files under tools/config which are now unused. New files are
added for the generator:

  - climate/opts.py is for collecting climate options
  - etc/climate/climate-config-generator.conf contains generator configs
  - etc/climate/README-climate.conf.txt describes how to generate a
    sample configuration file

In addition, a testenv:genconfig section is added to tox.ini and
settings for the oslo.config.opts entry point are added in setup.cfg.

[1] http://docs.openstack.org/developer/oslo.config/generator.html

Change-Id: Ifd352dee176310fd667d1523a85afd3ddf9ab5da
Closes-Bug: #1658908