Make DevStack create a freepool aggregate
Creating the Nova host aggregate used by Blazar as freepool for host
reservation was done automatically by the old extras.d plugin, but this
functionality was not migrated to the new DevStack plugin.

This commit creates the freepool before starting Blazar services. It
also uses the openstack client rather than the nova command.

Change-Id: Ifc6555855342ee274e01b1f1bc3ace913403d50a