Support host reservation scenario test
The existing Tempest scenario test in Blazar only supports instance
reservation. However, the Blazar service currently only supports host
reservations.

This patch adds a Tempest scenario test for host reservation. It checks
that Blazar correctly reserves a host for a reservation and that an
instance created with the reservation ID as scheduler hint is correctly
scheduled to the host.

Change-Id: Ibbf33ff6344bdc26e005e1fa4252c317fdbdfb90
Closes-Bug: #1659216