Add namespace release note
In O release, Blazar has moved all namespace from climate to blazar.
This requires users of Blazar to revise their config files because of
the namespace migration.

This patch adds a release note about upgrade and deprecation, that
climate commands and config values are supported in O but will be
removed in the P release.

Change-Id: I467b5ec65ad0bafee03590c875379663c18b1ece
Closes-Bug: #1658859
Closes-Bug: #1662734