Allow terminating leases at anytime
This patch enables lease termination at anytime even if the lease has
already started. It supports two ways of terminating a lease:

1. Update lease with "end_date" = "now"

Blazar updates the end time of the lease with the current time when it
accepts the request. Then, the lease will be terminated soon.

2. Delete lease

Blazar terminates and deletes the lease at the time when it accepts the
request.

Change-Id: Idacc4ad56863adba0c8ee14929a35bcc9336fe08
Implements: blueprint terminate-lease-at-anytime