Gather all operations for the reservations table into the manager
Records in the reservations table in the Blazar DB are created and
updated by both the manager and the resource plugin [1]. This results in
a lack of separation of concerns.

This patch migrates such operations from the resource plugin to the
manager. It makes the manager responsible for the reservations table,
and the resource plugin responsible only for resource specific tables
(e.g. the computehost_reservations table).

It also changes the resource_id field in the reservations table to hold
an ID of the record of resource-specific reservations tables, e.g.
computehost_reservations table for the physical-host plugin and
instance_reservations table for the new instance plugin [2].
An ID of resource itself which was held in the resource_id field is
migrated to the resource specific tables, so a new field 'aggregate_id'
is added into the computehost_rservations table.

The instance_reservations table of the new instance plugin is expected
to have resource IDs like a flavor ID, and aggregate ID, and instance
IDs. The ID of the record of the instance_reservations table is expected
to be held in the resource_id field of the reservations table.

[1] https://etherpad.openstack.org/p/blazar-resource-plugin
[2] I03315216c3a6203e088e914ffb9fedf1e672d732

Change-Id: Ia042020088af9e34e3430c44814cd69f4ed303fb