Use eventlet executor instead of blocking executor
The default executor of oslo.messaging is 'blocking'. The blocking
executor has been broken by the patch[1]. And the issue has been
published in oslo.messaging 5.25.0.

This patch changes the executor used by rpc_server in blazar-manager.
Additionally, in oslo team the blocking executor will be removed in
the future release.

1. https://review.openstack.org/#/c/463673/

Change-Id: I87dae619d4b89a1f2465d31a62d46de8d26f7512