Fix code for deleting instances when leases end
The code deleting instances when leases end runs Nova queries as the
Blazar service user. Because it was not using the optional all_tenants
parameter, the query would only return instances owned by the service
project.

This commit sets the all_tenants parameter, allowing the query to return
all existing instances on the host no matter which project started them.
It also extends the Tempest host reservation scenario to delete the
lease and verifies that the instance is correctly terminated.

Change-Id: I6fed0ca12505c94a709f8070dfc6e26d2093345f
Closes-Bug: #1699107