Prevent host reallocation for active reservations
Extending the length of a physical host reservation can lead Blazar to
reallocate the reservation to new hosts if some of the reserved hosts
are not available during the extended lease period. This can happen even
if the lease has already started, as long as there is no instance
running on any of the hosts to reallocate.

This can cause a race condition:

1. A lease starts
2. A user requests to extend their lease
3. The host plugin checks the availability of the reserved host. It
finds the reserved host (host A) is not available for the extended
period, but another host (host B) is available.
4. The host plugin checks if any instances are running on host A and
finds that there is no instance.
5. At the same time, the user launches an instance on host A.
6. The host plugin does not notice it, and proceeds to reallocate the
reservation from host A to host B. This causes a serious problem because
the instance is running on a host that does not belong to the
reservation anymore.

To solve this problem, the update_reservation method is changed to only
reallocate hosts if the reservation has not yet started.

Change-Id: I62a62cb2497b2c22de722ae71bd060f2f3935010
Closes-Bug: #1692805