Move the convert_requirements method to the utils namespace
The convert_requirements method is used by host_plugin and
instance_plugin. But this method is implemented as an instance method of
host_plugin. The instance_plugin can't use this method without an
instance of host_plugin.

This patch moves the convert_requirements method to the utils namespace.
All plugins which query the db data can use this method.

Partially implements: blueprint new-instance-reservation
Change-Id: Ib365a794b45a1a1bf0c1b0a8a750171d44d25b8a