Fix eventlet monkey patching
Eventlet monkey patching needs to be done as early as possible,
otherwise previous library imports are not patched correctly. This fixes
issues with concurrent requests receiving a 500 Internal Server Error
with a traceback including the following message:

    AttributeError: 'function' object has no attribute 'get'

This commit also adds a new Tempest scenario to test for this issue by
running concurrent requests against the Blazar service.

Change-Id: I9be2b6787a151d54ca9cce5c91fa9e11e727ce0b