Fix host filtering by extra capability
When filtering nodes by extra capabilities, the host_get_all_by_queries
method would also select hosts that do not have the extra capability. A
request for nodes with resource_properties set to '["=", "$foo", "bar"]'
would select all nodes with have an extra capability with name "foo" and
value "bar", but also all nodes which do not have an extra capability
named "foo".

This commit ensure that hosts which do not have a matching extra
capability name are filtered out of the result.

Change-Id: Id1882780d51ebdf417541bb5d6764c590097c301
Closes-Bug: #1702745