Fix race condition between lease creation and event thread
If a lease is created to start immediately, there is a possible race
condition between the lease creation function (create_lease() in
blazar/manager/service.py) and the manager thread running lease events.

The race condition happens in the following fashion:

1. create_lease() calls db_api.lease_create(lease_values). The
lease_values dict contains events values, which means both lease and
associated events are created. However, associated reservations are not
yet created at this stage.
2. The manager thread running events is triggered by its periodic timer.
It finds that the start_lease event is ready to run and executes
_basic_action(). Since there is no reservation yet, _basic_action() only
marks the event status as 'DONE'.
3. create_lease() finishes running by creating reservations associated
with the lease.
4. Reservations stay in "pending" status forever because the start_lease
event has already run.

This is resolved by separating event creation in the database from the
lease creation and moving it after reservation creation is done.

Change-Id: Ia954461a682b4c5eef6dd5f3d5b779650ce404aa
Closes-Bug: #1704359