Skip on_end action if reservation has already ended
When a reservation reaches its lease end time, the on_end action is run
via the end_lease event and deletes the corresponding pool after
terminating instances. However, it is also run when calling
delete_lease, regardless of the reservation status.

This leads to the following scenario. A reservation reaches its lease
end time and on_end is run, which deletes the pool. Then, a user deletes
the corresponding lease, which results in on_end being run again for the
same reservation. Logging at DEBUG level will show that the second
on_end call tries to access the pool, which has already been deleted.

    RESP BODY: {"itemNotFound": {"message": "Aggregate 2 could not be found.", "code": 404}}

This commit skips the on_end action if the host reservation is already
marked as deleted (which actually means it has ended, not that the lease
was deleted by the user).

Change-Id: I8d1e4cae5213b825fd81d676ecca8527d6b910bb