Return a MissingParameter error when required parameters are missing
The instance reservation plugin was failing to raise a MissingParameter
exception in its parameter validation method. It resulted in users
receiving a "500 Internal Server Error" when trying to create instance
reservations with insufficient parameters.

This patch changes the plugin to correctly raise a MissingParameter
exception in the validation method, which results in a MissingParameter
error with a 400 status code being returned to the user.

Change-Id: I2217eb60496e4f564b0e5181445e9ac919c70d3d