Set user_id when creating leases
The user_id field in leases is set from ctx.user_id, which is always
None because it is not defined in trusts.create_ctx_from_trust().
Nevertheless, we cannot use this value since we are in the trustee
context, which is the blazar user.

This patch modifies the API handler so that the real user_id is included
in the lease_values dict passed to create_lease(). Ideally, the context
would be extended to include a field referring to the trustor user.

Change-Id: If0ee6efcd30dc56dfd78fbfe566517a8d8aa6ba7
Closes-Bug: #1709103