Add support for multi-region deployments
In multi-region deployments using a shared Keystone, with each region
using its own Blazar and Nova services, Blazar could not be configured
to be part of a specific region. It would use the first endpoint found
in the service catalog. As a result, Blazar in RegionTwo would try to
manage Nova hosts and aggregates in RegionOne.

This patch adds a new configuration option called os_region_name in the
[DEFAULT] section. If set, Blazar looks up the right endpoints for its
region. The DevStack plugin is also updated to use the correct region
and Keystone host (which is otherwise set to the default of 127.0.0.1).

Change-Id: I0e548103dbd3d50af366ad5c19c98724e22bdc3d
Closes-Bug: #1694754