Add a status module
This patch adds a new status module which defines statuses of a lease,
reservation and event. It also provides basic features for managing
these statuses.

The LeaseStatus class provides a lease_status method which is used as a
decorator for the *-lease method of the ManagerService class.
A status of a lease is checked and updated before/after executing
*-lease methods by decorating these methods. This patch decorates some
*-lease methods with it.

This patch also replaces status literals with members of the *Status
classes, e.g. the 'pending' literal is replaced by
ReservationStatus.PENDING.

Implements: blueprint state-machine
Change-Id: I014aba977e5203bb72e055d08716f84ff76015cc