Add a resource monitoring feature into the manager
This patch adds a new monitor module which provides the following two
types of monitor class:

* NotificationMonitor: It detects resource failures by subscribing
  notifications and handles these failures.

* PollingMonitor: It detects resource failures by periodically calling
  APIs of other services like Nova and handles these failures.

The NotificationMonitor listens to notification messages by using the
notification listener of the oslo_messaging[1]. Topics to subscribe and
sets of an event type and a callback are provided by resource monitor
plugins.

The PollingMonitor periodically calls APIs of other services. The actual
poll() method and a polling interval are provided by resource monitor
plugins.

This patch also adds a base resource monitor plugin (the
BaseMonitorPlugin class). Actual implementations that inherit it will
follows this patch. Note that part of a monitoring plugin for the host
plugin is implemented as a sample.

[1] https://docs.openstack.org/oslo.messaging/latest/reference/notification_listener.html

Partially Implements: blueprint resource-monitoring
Change-Id: Ib3191e8ba67a784c5b5d009519161a847e94d55a