Support host recovery
This patch adds a host recovery feature to the host monitor plugin.
When recovery of a failed host is detected, it updates the reservable
field of the host to True.

Note that this status change does not trigger healing of degraded
leases. The recovered hosts can be re-reserved by calling Create/Update
lease request.

Partially Implements: blueprint resource-monitoring
Change-Id: Ie6767902a9d7dc60b3fa6c525a9a1270578ea49b