Use waiter method to check server status
The status of an instance could be 'ACTIVE' right after a lease ends,
since the task of deleting the instance is still in progress.

This patch replaces the assertion to check whether the instance is
terminated with a waiter method. The method periodically checks for the
existance of the specified instance until the instance is deleted.

Change-Id: I22a48f9b7b32a9e7df9e9319a395f91e648ba5a7