Retry event actions
An event action fails if the target lease is at a transitional status,
e.g., trying to start a lease while updating the lease and it fails.

This patch adds capabilities for retrying such an event action.
A new configuration variable 'event_max_retries' is added with this
change.

Change-Id: I34c8f1816660e4aa92404459d8908bf707785d0f
Implements: blueprint event-retry