Replace db_api with sql query in migration script
The db migration scripts are executed one by one. On the other
hands, db_api always expects the up-to-date DB schema is defined
in the Blazar DB. If db_api methods are executed in the middle
of db migrations, the mis-match between the real schema and model
definition causes migration error.

This patch replaces db_api methods in migration scripts with
sql queries. It prevents migration scrpits from the error.

Change-Id: I0509b9c7dadd2df5e70360fe4673f8a71a4c5f3a
Closes-Bug: #1757076