Handle availability zone info in compute host APIs
The create host API stores availability zone information into the
computehosts tables. The information comes from Nova's availability
zone API.

Cloud users can specify the availability zone information by following
query in the host reservation and the instance reservation.

- The host reservation
  hypervisor_properties=["==", "$availability_zone", "zone1"]

- The instance reservation
  resource_properties=["==", "$availability_zone", "zone1"]

The az_aware config is a flag to store the infomation or not.
If a cloud operator doesn't prefer to store the information, set the
flag to False.

Implements: blueprint multi-freepools
Change-Id: I26a7e635660eec54b0bef9754ef46aa5af7fe60e