Prevent the host update API from updating in-use extra capabilities
The host update API allows cloud admins to update extra capabilities
even though the host reservation or the instance reservation requested
the capabilities. It causes a mismatch between requested
resource_properties of reservations and extra_capabilities
registered in Blazar.

This patch prevents the API from updating extra capabilities if the key
is referred to in resource_properties.

Change-Id: I78a9dc7f448a29f31fb95432cf4eee66c865aa70
Implements: blueprint flavors-extra-specs