Ignore keys unsupported by BlazarContext in its creation
BlazarContext class raises an error when the class is instantiated
with unspported keys. The class is used as a context for RPC and
the RPC is implemented over oslo.messaging. If oslo.messaging
automatically adds a new key for the context, the RPC execution
fails since the new key is unexpected by BlazarContext and it
raises an error.

This patch relaxes the limitations in BlazarContext. If unexpected
keys are passed for the class instantiation, the class ignores the
keys.

Change-Id: Ibdca6eead0ee726bc10a07ee43e4ef10110d23e2
Closes-Bug: #1775363