Change response status code when start_date and end_date are invalid
The create lease API and the update lease API return a 403 response
status code when start_date and end_date have an invalid time order.
A 403 response status code means the API request is Not Authorized,
which can confuse the client.

This patch changes blazar-api to return a 400 response status code when
the two dates have an invalid time order.

Change-Id: I5603212b8c70388aa7d2da883ec0c62f564807e5