Add cleaning time
This patch adds a margin between the end of a lease and the start of the
next lease. The length of the margin can be configured with the new
configuration option 'cleaning_time'.

Note that it does not actually change start/end_lease time, but just
takes into account the margin when looking for available resources.

Implements: blueprint cleaning-time-allowance
Change-Id: I7de05a65c2221c1eb7c5b2f65bec58d6a48b2085