Change invalid input error code to 400
Now blazar instance plugin supports only ``affinity = False`` option.
This is because we have no scheme to check if the resource capability
is exceeded when mutiple instances land on the same compute host.

Currently we raise ``500`` (internal server) error when
``affinity = True`` is set by users. However, this error should be
``400`` (bad request error) to explicitly inform users that we don't
support it so far.

This patch fixes it to return ``400`` error using the
MalformedParameter() exception. A releasenote for this fix will be
added in a following patch.

Change-Id: Iea98a4d7c448470148b58982281b5d1e9f1892b5