Return 400 Bad Request for extra capability keys longer than 64 chars
Length of extra capability keys is limited to 64 characters in the DB
schema. However, the host create API can create a host based on a
request that has a longer extra capability key, only to return a 500
Internal Server Error.

With this patch, the API checks the key length before creating a host
and returns a 400 Bad Request status code if it is too long.

Change-Id: I2cf4ede787e0f65385376832ce410d694df094ff
Closes-Bug: #1709639