Correct the project domain option for authorization
In blazar configuration, we have the following options:

* os_admin_user_domain_name
* os_admin_project_domain_name

They are used for Keystone authentication, but 
``os_admin_project_domain_name`` in the configuration file was not
reflected in Blazar. This is because the Blazar service used
``os_admin_user_domain_name" for both *project* and *user* domain name.

This patch fixes the bug and adds related tests.

Change-Id: Ifbca9f622afba3aa9f41296fff2d34afce22da0a
Closes-Bug: #1779660