Set Tempest concurrency to one
We are going to add API tests using Gabbi to the blazar-tempest-plugin.
But if the Gabbi tests are executed by different workers, their order
gets violated.

This patch changes Tempest concurrency to 1 to ensure the Gabbi tests
are executed in the order specified in the YAML file. This can increase
test execution time, but it should be negligible.

Change-Id: Id4dc3cc8640e8f902f96130a85e50008b5e8b96d