Raise database exception instead of returning None
In case of database errors when creating a new compute host, the host
plugin would not raise the database exception after removing the host
from the freepool, but would instead return None. This is not helpful
for operators who get the following message from the CLI:

    'NoneType' object has no attribute 'items'

This patch raises the database exception again, which results in error
messages such as:

    ERROR: BlazarDBDuplicateEntry: A database error occurred

The code returning None is removed as it is now unneeded.

Change-Id: If4240b5a191e2a4e3acee120c53517af029d96b7