Allow terminating a lease in ERROR state
This patch adds a valid state transition from ERROR to TERMINATING,
which should allow a lease in the ERROR state to at least attempt to
clean itself up upon lease end. If this state transition is not allowed,
leases in the ERROR state effectively will not be cleaned up and their
hosts will stay in the corresponding aggregate instead of being returned
to the freepool.

This is compounded by the fact that leases can easily go into ERROR
state because of bug https://bugs.launchpad.net/blazar/+bug/1786031.

Change-Id: I1fbd922d921b7a643379c93ca6b2ebb4edde9c7a
Closes-Bug: #1793560