Adds subnormal path for placement client (1)
We start to support placement api from Stein, so there are cases where
the child resource provider of a host doesn't exist if the host has
been created in blazar before Rocky release.

This patch changes the placement client code to be aware of the cases.
Namely, in cases the resource provider is not found,

- Change get_resource_provider() to return None
- Change delete_reservation_provider() to skip the operation
- Change update_reservation_inventory() to create the resource provider
- Change delete_reservation_inventory() to raise an error

Change-Id: I82abbc73f69f4036f60b7c549c1cbd0c0b91a130
Blueprint: placement-api