Adds subnormal path for placement client (2)
We start to support placement api from Stein, so there are cases where
the resource classs of the reservation id doesn't exist if the
reservation been created before Rocky release.

This patch changes the placement client code to be aware of the cases.
Namely, in cases the resource class is not found, this patch changes
delete_reservation_inventory() and delete_reservation_class() to skip
the operation.

Change-Id: I5c4a4b8c90f237fa392c2a3ae3ffc9be64df7d64
Blueprint: placement-api