Add allocated hosts to aggregate on active reservation update
When an active reservation was updated to increase the number of
reserved hosts, new hosts allocations were created, but the
corresponding Nova hosts were not added to the host aggregate. This
prevented users from deploying instances on the newly allocated hosts.

This patch ensures newly allocated hosts are added to the aggregate when
updating active reservations.

Change-Id: I11904934319665c9e75558928c87fc92d856325d
Closes-Bug: #1800463